requirements.txt
mem0ai==0.1.20
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
chromadb==0.4.18
